training program with smart method, fourth meeting
the Automatic and not automatic robot starting algorithm. 
In this algorithm, you will find the way of controlling the robot in an automatic or not automatic way. 
The automatic will be just by the robot, but if you press one of the buttons in the browser connected with the robot you will be controlling.